Update _the_general_epistle_of_barnabas.md
https://archive.org/details/theepistlesofscl00hooluoft/page/n149/mode/2up
https://archive.org/details/theepistlesofscl00hooluoft/page/n151/mode/2up
https://archive.org/details/theepistlesofscl00hooluoft/page/n153/mode/2up
https://archive.org/details/theepistlesofscl00hooluoft/page/n155/mode/2up
https://archive.org/details/theepistlesofscl00hooluoft/page/n155/mode/2up
https://archive.org/details/theepistlesofscl00hooluoft/page/n157/mode/2up
https://archive.org/details/theepistlesofscl00hooluoft/page/n159/mode/2up

Enoch References: 

80:2 And in the days of the sinners the years shall be shortened, And their seed shall be tardy on their lands and fields, And all things on the earth shall alter, And shall not appear in their time: And the rain shall be kept back And the heaven shall withhold (it).

89:67 And the lions and tigers eat and devoured the greater part of those sheep, and the wild boars ate along with them; and they burnt that tower and demolished that house.

There is a second book of Ezekiel, where it mentions blood flowing from a tree (this part is missing, but the first part of that verse is still intact) Look at Frag. 2 https://ancient-scriptures.fandom.com/wiki/The_Second_Book_Of_Ezekiel

2 Esdras 12: (vision with the eagle)

Verses regarding food as symbolism for doctrine (that I can think of)
Deuteronomy 30:14 
Jeremiah 15:16
Ezekiel 3:3
Revelation 10:10
John 4:34
1 Corinthians 3:2

Side note: In Mark 7:19, it says this 
18 And he saith unto them, Are ye so without understanding also? Do ye not perceive, that whatsoever thing from without entereth into the man, it cannot defile him;



19 Because it entereth not into his heart, but into the belly, and goeth out into the draught, purging all meats?

In other translation it says (In saying this, Jesus declared all foods clean.) Since it is in parentheses, it could be assumed it was added. I only found that KJV and NKJV don't have this. (Someone took liberties...)

Might be helpful to mention Romans 14